fix/cleanup/adjustment: flirting and dating restrictions
condensed the code that determined whether a cat could be flirted with or taken on a date into a new function located within cats.py, is_dateable.

is_dateable is quite similar to is_potential_mate, with a few differences to allow for player agency and to give freedom in editing the function without messing with is_potential_mate.

previously, flirting was reliant on rank rather than age, which could lead to uncomfortable situations such as a 6 moon old being able to flirt with a 15 moon old if they were both apprentices.

on the other hand, dating used is_potential_mate, which would lead to odd situations such as being able to flirt with cats at 12 moons old but not take them on dates until 14 moons, even if they were your mate.

now, both dating and flirting rely on is_dateable. if a cat is between 6-11 moons in age they should only be able to flirt with other 6-11 moon olds now. once a cat reaches 12 moons old, they can flirt with or date any cats that fall within the age range denoted in game_config, provided of course that those cats are older than 11 moons,  not relatives, and not former/current mentors/apprentices.

personally i would suggest maybe taking off the restriction for dating that the cat must not be in game.patrolled because sometimes you wanna take your spouse on a date after work, but i have left that as it is for right now.